Fix warnings
