+++ src/main/webapp/app/src/App.jsx
@@ -2,8 +2,8 @@ import * as React from 'react'
 import { useEffect, useState } from 'react'
 import CssBaseline from '@mui/material/CssBaseline'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import Footer from './components/core/Footer'
-import NavBar from './components/core/NavBar'
+import Footer from './components/footer/Footer'
+import NavBar from './components/navbar/NavBar'
 import Container from '@mui/material/Container'
 import AppContext, { defaultAppContext } from './core/AppContext'
 import { ErrorBoundary } from 'react-error-boundary'
