+++ src/main/webapp/app/src/components/navbar/NavBar.jsx
@@ -14,9 +14,9 @@ import * as React from 'react'
 import { useContext, useState } from 'react'
 import logo from './logo_banner.png'
 import { Link, useNavigate } from 'react-router-dom'
-import LogInModal from '../authentication/LogInModal'
-import ForgetPasswordModel from '../authentication/ForgetPasswordModal'
-import SignUpModal from '../authentication/SignUpModal'
+import LogInModal from './authentication/LogInModal'
+import ForgetPasswordModel from './authentication/ForgetPasswordModal'
+import SignUpModal from './authentication/SignUpModal'
 import AppContext, { defaultAppContext } from '../../core/AppContext'
 import ajax from '../../core/ajax'
