+++ src/main/webapp/app/src/components/navbar/authentication/LogInModal.jsx
@@ -1,4 +1,5 @@
 import * as React from 'react'
+import { useContext } from 'react'
 import Backdrop from '@mui/material/Backdrop'
 import Box from '@mui/material/Box'
 import Modal from '@mui/material/Modal'
@@ -8,12 +9,16 @@ import Typography from '@mui/material/Typography'
 import Avatar from '@mui/material/Avatar'
 import CssBaseline from '@mui/material/CssBaseline'
 import TextField from '@mui/material/TextField'
+import FormControlLabel from '@mui/material/FormControlLabel'
+import Checkbox from '@mui/material/Checkbox'
 import Link from '@mui/material/Link'
 import Grid from '@mui/material/Grid'
 import LockOutlinedIcon from '@mui/icons-material/LockOutlined'
 import Container from '@mui/material/Container'
-// import ReCAPTCHA from "react-google-recaptcha";
-// need domain to use
+import RetryAlert from './RetryAlert'
+import ajax from '../../../core/ajax'
+import AppContext from '../../../core/AppContext'
+import { useNavigate } from 'react-router-dom'
 
 const style = {
   position: 'absolute',
@@ -27,31 +32,43 @@ const style = {
   p: 4
 }
 
-const ForgetPasswordModal = (prop) => {
-  const { visible, setVisible } = prop
-  const handleClose = () => setVisible('')
+const LogInModal = ({ visible, setVisible }) => {
+  const navigate = useNavigate()
+  const { context, setContext } = useContext(AppContext)
 
-  const handleSwitchtoSignIn = () => {
-    setVisible('signin')
+  const handleClose = () => {
+    setRetryAlert(false)
+    setVisible('')
+  }
+
+  const [retryAlert, setRetryAlert] = React.useState(false)
+
+  const handleSwitchtoForgetPassword = () => {
+    setVisible('forgetPassword')
   }
 
   const handleSwitchtoSignUp = () => {
     setVisible('signup')
   }
 
-  const handleSubmit = (event) => {
+  const handleSubmit = async (event) => {
     event.preventDefault()
-    const data = new FormData(event.currentTarget)
-    console.log({
-      email: data.get('email'),
-      password: data.get('password')
-    })
+    const formData = new window.FormData(event.currentTarget)
+    try {
+      const user = await ajax('POST', formData)('/login')
+      setContext({ ...context, user })
+      handleClose()
+      navigate('/dashboard')
+    } catch (e) {
+      console.log(e) // FIXME
+      setRetryAlert(true)
+    }
   }
 
   return (
     <Modal
       aria-labelledby='transition-modal-title'
-      open={visible === 'forgetPassword'}
+      open={visible === 'signin'}
       onClose={handleClose}
       // closeAfterTransition
       BackdropComponent={Backdrop}
@@ -60,9 +77,9 @@ const ForgetPasswordModal = (prop) => {
       }}
       sx={{ width: '70%', left: '15%' }}
     >
-      <Fade in={visible === 'forgetPassword'}>
+      <Fade in={visible === 'signin'}>
         <Box sx={style}>
-          {/* white box to hold form */}
+          {retryAlert && <RetryAlert />}
           <Container component='main' maxWidth='xs'>
             <CssBaseline />
             <Box
@@ -80,7 +97,7 @@ const ForgetPasswordModal = (prop) => {
                 component='h1'
                 variant='h5'
               >
-                Reset Password
+                Sign in
               </Typography>
               <Box
                 component='form'
@@ -98,22 +115,35 @@ const ForgetPasswordModal = (prop) => {
                   autoComplete='email'
                   autoFocus
                 />
-
+                <TextField
+                  margin='normal'
+                  required
+                  fullWidth
+                  name='password'
+                  label='Password'
+                  type='password'
+                  id='password'
+                  autoComplete='current-password'
+                />
+                <FormControlLabel
+                  control={<Checkbox value='remember' color='primary' />}
+                  label='Remember me'
+                />
                 <Button
                   type='submit'
                   fullWidth
                   variant='contained'
                   sx={{ mt: 2, mb: 2 }}
                 >
-                  Send Recovery Email
+                  Sign In
                 </Button>
                 <Grid container>
                   <Grid item xs>
                     <Link
                       variant='body2'
-                      onClick={() => handleSwitchtoSignIn()}
+                      onClick={() => handleSwitchtoForgetPassword()}
                     >
-                      Sign In
+                      Forgot password?
                     </Link>
                   </Grid>
                   <Grid item>
@@ -121,7 +151,7 @@ const ForgetPasswordModal = (prop) => {
                       variant='body2'
                       onClick={() => handleSwitchtoSignUp()}
                     >
-                      Don't have an account? Sign Up
+                      {'Don\'t have an account? Sign Up'}
                     </Link>
                   </Grid>
                 </Grid>
@@ -134,4 +164,4 @@ const ForgetPasswordModal = (prop) => {
   )
 }
 
-export default ForgetPasswordModal
+export default LogInModal
