+++ src/main/webapp/app/src/components/schedule/ScheduleEventFooter.js
