+++ src/main/webapp/app/src/pages/Wizard.jsx
@@ -7,6 +7,7 @@ import Button from '@mui/material/Button'
 import LoadingButton from '@mui/lab/LoadingButton'
 import { Link } from 'react-router-dom'
 
+// FIXME remove?
 const Wizard = () => {
   const [rosterBtnLoading, setRosterBtnLoading] = useState(false)
